fix: dialyzer errors due to opaque types
Prior to this commit there were several places where we were
trying to reference properties on opaque types from outside
of the corresponding modules. This caused dialyzer errors.
This commit refactors the classes a bit to reduce the number
of places where a class has any need to try to dig into
the opaque types, and in the places where it is still necessary,
uses accessor functions rather than referencing the properties
directly.