feat: standardize the response types
Change all methods that return 'success' to instead return the more
idiomatic 'ok'.

Change all methods that return an atom when they succeed to return a
struct as well, so we can add fields in the future without breaking
customer code.

Change get to return a struct containing the value instead of the bare
value.