chore: make set ttl optional and misc fixes
Make CacheClient.set() ttl_seconds optional to match the other functions
that take a ttl.

Fix a whitespace issue that was causing a warning on @enforce_keys in
Configuration.

Make the integration tests create and delete the test cache.

Make the github action use a unique cache name for each test run.

Fix a validation error that occurs in elixir 1.15.

Add a missing case to get_rank in which the sorted set exists but the
desired value does not.