feat: Add doc code snippets and readme generation (#43)
* feat: Add doc code snippets and readme generation

Add code snippets for the public dev docs

Add a readme template and update the actions to generate and verify
the readme.

Update the examples to gomomento 0.4.0 and make a few fixes to let them
build.

Change :hit to :ok in the CacheClient.get() response. This is not yet
reflected in the examples/snippets. They will need to be updated after
the next release.