feat: support numbers for ttls/scores
This commit modifies the places where we were requiring floats
as inputs to accept numbers instead. This mostly impacts TTLs,
and just allows users to pass in an integer TTL if they prefer.
Also allows users to pass in integers for sorted set scores,
though we will still use floats for the outputs (response classes).