feat: use keyword list for cache client factory function
This commit tweaks the cache client factory function to support
a keyword list for its arguments. This allows callers to pass
the arguments as named args, and change the order if they like.
It also makes it a little more straightforward for us to provide
default values for any of the options in the future.

Also adds a `create` variant in addition to our `create!`, so that
users can opt in to pattern matching for errors if they prefer.