chore: Make error handling and validation consistent (#60)
Bring the older functions in line with the later ones that have lighter
explicit validation in favor of transformed protobuf EncodeErrors.

Ensure that the ttl handling works if using a collection ttl with nil
ttl_seconds and that an error is returned for a malformed collection ttl
or ttl.

Combine the functions that previously had a separate send function.

Remove now unused validation methods.

Remove the now unused cache_client_test. The validation tests now
require a real grpc client.